Upgrading @angular & @angular/material versions
