+++ ontimize/components/input/listpicker/o-list-picker.component.ts
@@ -19,13 +19,12 @@ import { InputConverter } from '../../../decorators';
 import { OFormComponent } from '../../form/o-form.component';
 import { OSearchInputModule } from '../../search-input/o-search-input.component';
 import { OFormValue } from '../../form/OFormValue';
-import { Util } from '../../../utils';
 import { ODialogModule } from '../../dialog/o-dialog.component';
 
 import { OListPickerDialogComponent } from './o-list-picker-dialog.component';
 import { OTranslateModule } from '../../../pipes/o-translate.pipe';
 
-import { OFormDataComponent } from '../../o-form-data-component.class';
+import { OFormServiceComponent } from '../../o-form-service-component.class';
 
 export const DEFAULT_INPUTS_O_LIST_PICKER = [
   'oattr: attr',
@@ -81,36 +80,16 @@ export const DEFAULT_OUTPUTS_O_LIST_PICKER = [
   ],
   encapsulation: ViewEncapsulation.None
 })
-export class OListPickerComponent extends OFormDataComponent implements OnInit {
+export class OListPickerComponent extends OFormServiceComponent implements OnInit {
 
   public static DEFAULT_INPUTS_O_LIST_PICKER = DEFAULT_INPUTS_O_LIST_PICKER;
   public static DEFAULT_OUTPUTS_O_LIST_PICKER = DEFAULT_OUTPUTS_O_LIST_PICKER;
 
   /* Inputs */
-  protected staticData: Array<any>;
-
-  protected entity: string;
-  protected service: string;
-  protected columns: string;
-  protected valueColumn: string;
-  protected parentKeys: string;
-  protected visibleColumns: string;
-  protected descriptionColumns: string;
-
   @InputConverter()
   protected filter: boolean = true;
   protected separator: string;
-
-  @InputConverter()
-  protected queryOnInit: boolean = true;
-  @InputConverter()
-  protected queryOnBind: boolean = false;
   /* End inputs */
-
-  protected dataArray: any[] = [];
-  protected colArray: string[] = [];
-  protected visibleColArray: string[] = [];
-  protected descriptionColArray: string[] = [];
 
   @ViewChild('dialog')
   protected dialog: MdDialog;
@@ -120,12 +99,6 @@ export class OListPickerComponent extends OFormDataComponent implements OnInit {
 
   public onChange: EventEmitter<Object> = new EventEmitter<Object>();
 
-  protected dataService: OntimizeService;
-  protected cacheQueried: boolean = false;
-  protected _pKeysEquiv = {};
-  protected _formDataSubcribe;
-  protected _currentIndex;
-
   constructor(
     @Optional() @Inject(forwardRef(() => OFormComponent)) form: OFormComponent,
     elRef: ElementRef,
@@ -134,43 +107,7 @@ export class OListPickerComponent extends OFormDataComponent implements OnInit {
   }
 
   ngOnInit(): any {
-    this.cacheQueried = false;
     this.initialize();
-
-    this.colArray = Util.parseArray(this.columns);
-
-    this.visibleColArray = Util.parseArray(this.visibleColumns);
-    if (Util.isArrayEmpty(this.visibleColArray)) {
-      //It is necessary to assing value to visibleColumns to propagate the parameter to dialog component.
-      this.visibleColumns = this.columns;
-      this.visibleColArray = this.colArray;
-    }
-
-    this.descriptionColArray = Util.parseArray(this.descriptionColumns);
-    if (Util.isArrayEmpty(this.descriptionColArray)) {
-      this.descriptionColArray = this.visibleColArray;
-    }
-
-    let pkArray = Util.parseArray(this.parentKeys);
-    this._pKeysEquiv = Util.parseParentKeysEquivalences(pkArray);
-
-    if (this.form) {
-      var self = this;
-      if (self.queryOnBind) {
-        this._formDataSubcribe = this.form.onFormDataLoaded.subscribe(data => {
-          self.onFormDataBind(data);
-        });
-      }
-    }
-
-    if (this.staticData) {
-      this.queryOnBind = false;
-      this.queryOnInit = false;
-      this.setDataArray(this.staticData);
-    } else {
-      this.configureService();
-    }
-
   }
 
   ensureOFormValue(value: any) {
@@ -179,30 +116,13 @@ export class OListPickerComponent extends OFormDataComponent implements OnInit {
     } else if (value && !(value instanceof OFormValue)) {
       this.value = new OFormValue(value);
     } else {
-      this.value = new OFormValue(undefined);
+      this.value = new OFormValue(this.defaultValue);
     }
     this.syncDataIndex();
   }
 
-  configureService() {
-    this.dataService = this.injector.get(OntimizeService);
-
-    if (Util.isDataService(this.dataService)) {
-      let serviceCfg = this.dataService.getDefaultServiceConfiguration(this.service);
-      if (this.entity) {
-        serviceCfg['entity'] = this.entity;
-      }
-      this.dataService.configureService(serviceCfg);
-    }
-
-  }
-
   ngOnDestroy() {
     this.destroy();
-
-    if (this._formDataSubcribe) {
-      this._formDataSubcribe.unsubscribe();
-    }
   }
 
   ngAfterViewInit(): void {
@@ -211,78 +131,6 @@ export class OListPickerComponent extends OFormDataComponent implements OnInit {
     } else if (this.queryOnBind) {
       //TODO do it better. When changing tabs it is necessary to invoke new query
       this.syncDataIndex();
-    }
-  }
-
-  onFormDataBind(bindedData: Object) {
-    let filter = {};
-    let keys = Object.keys(this._pKeysEquiv);
-    if (keys && keys.length > 0 && bindedData) {
-      keys.forEach(item => {
-        let value = bindedData[item];
-        if (value) {
-          filter[this._pKeysEquiv[item]] = value;
-        }
-      });
-    }
-    this.queryData(filter);
-  }
-
-  queryData(filter?: Object): void {
-    let kv = {};
-    let self = this;
-    if (filter) {
-      kv = filter;
-    }
-    if (this.dataService === undefined) {
-      console.warn('No service configured! aborting query');
-      return;
-    }
-    this.dataService.query(kv, this.colArray, this.entity)
-      .subscribe(
-      (res: any) => {
-        if (res.code === 0) {
-          self.cacheQueried = true;
-          self.setDataArray(res.data);
-        } else {
-          console.log('error');
-        }
-      },
-      (error: string) => {
-        console.log(error);
-      }
-      );
-  }
-
-  setDataArray(data: any): void {
-    if (Util.isArray(data)) {
-      this.dataArray = data;
-      this.syncDataIndex();
-    } else if (Util.isObject(data)) {
-      this.dataArray = [data];
-    } else {
-      console.warn('ListPicker has received not supported service data. Supported data are Array or Object');
-      this.dataArray = [];
-    }
-  }
-
-  syncDataIndex() {
-    this._currentIndex = undefined;
-    if (this.value && this.value.value && this.dataArray) {
-      let self = this;
-      this.dataArray.forEach((item, index) => {
-        if (item[self.valueColumn] === this.value.value) {
-          self._currentIndex = index;
-        }
-      });
-
-      if (this._currentIndex === undefined) {
-        if (this.queryOnBind && this.dataArray && this.dataArray.length === 0
-          && !this.cacheQueried && !this.isEmpty()) {
-          this.queryData();
-        }
-        return;
-      }
     }
   }
